commit2

+++ src/main/java/com/example/Demo/Repositories/CateRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CateRepository extends JpaRepository<Category,Long> {
+public interface CateRepository extends JpaRepository<Category,Integer> {
     List<Category> findByCateName(String trim);
 }
